requirements.txt
streamlit
spacy
pdfminer.six
python-docx
plotly
matplotlib
scikit-learn
pandas
